Deleted old DNS Server and DNS Resolved libraries to favor the newer DNSLib version which will handle both A and TXT records
